+++ source/PhoneInputWithCountry.js
@@ -710,7 +710,7 @@ PhoneNumberInput.propTypes = {
 	/**
 	 * Country `<select/>` component props.
 	 * Along with the usual DOM properties such as `aria-label` and `tabIndex`,
-	 * some custom properties are supported, such as `arrowComponent`.
+	 * some custom properties are supported, such as `arrowComponent` and `unicodeFlags`.
 	 */
 	countrySelectProps: PropTypes.object,